Move back to main code being in audiblez.py (avoid a refactoring)
